[dev] lite: merge the two gradient accumulation test files
They shared their fixtures and covered the same contract from two angles --
Transformer Engine's fusion and lite's own accumulating paths -- so they now
live in one file.
