+++ experimental/lite/tests/unit/primitive/parallel/test_main_grad_accumulation_unit.py
@@ -6,6 +6,7 @@ from __future__ import annotations
 import pytest
 import torch
 
+from megatron.lite.primitive.optimizers.megatron_wrap import _enable_wgrad_accumulation_fusion
 from megatron.lite.primitive.parallel.linear import (
     AccumulatingLinear,
     _EmbeddingAccumulatingIntoMainGrad,
@@ -115,3 +116,79 @@ def test_embedding_reports_a_gradient_for_ddp() -> None:
 
     assert weight.grad_added_to_main_grad is True
     assert weight.grad is not None, "DDP would assert on a missing gradient"
+
+
+
+
+class _FakeTELinear(torch.nn.Module):
+    """Stands in for a TE linear: carries the flag and a weight, nothing else."""
+
+    def __init__(self, with_main_grad: bool) -> None:
+        super().__init__()
+        self.fuse_wgrad_accumulation = False
+        self.weight = torch.nn.Parameter(torch.zeros(2, 2))
+        if with_main_grad:
+            self.weight.main_grad = torch.zeros(2, 2, dtype=torch.float32)
+
+
+class _Plain(torch.nn.Module):
+    """A module with a weight but no TE flag -- must not grow one."""
+
+    def __init__(self) -> None:
+        super().__init__()
+        self.weight = torch.nn.Parameter(torch.zeros(2, 2))
+        self.weight.main_grad = torch.zeros(2, 2, dtype=torch.float32)
+
+
+def test_only_switches_modules_whose_weights_have_main_grad() -> None:
+    chunk = torch.nn.Module()
+    chunk.ready = _FakeTELinear(with_main_grad=True)
+    chunk.not_ready = _FakeTELinear(with_main_grad=False)
+    chunk.plain = _Plain()
+
+    assert _enable_wgrad_accumulation_fusion([chunk]) == 1
+    assert chunk.ready.fuse_wgrad_accumulation is True
+    # No main_grad to accumulate into: TE would dereference a buffer that the
+    # optimizer never allocated, so this one has to stay off.
+    assert chunk.not_ready.fuse_wgrad_accumulation is False
+    assert not hasattr(chunk.plain, "fuse_wgrad_accumulation")
+
+
+def test_is_idempotent_and_reports_nothing_switched_on_a_second_pass() -> None:
+    """Re-wrapping must not double-count; the caller reads the return value."""
+    chunk = torch.nn.Module()
+    chunk.ready = _FakeTELinear(with_main_grad=True)
+    assert _enable_wgrad_accumulation_fusion([chunk]) == 1
+    assert _enable_wgrad_accumulation_fusion([chunk]) == 0
+
+
+@pytest.mark.gpus(1)
+def test_fused_accumulation_matches_the_add_it_replaces() -> None:
+    """Same accumulated gradient as ``main_grad.add_(grad)``, over two microbatches."""
+    import transformer_engine.pytorch as te
+
+    torch.manual_seed(0)
+    shape = (64, 64)
+    inputs = [torch.randn(32, 64, device="cuda", dtype=torch.bfloat16) for _ in range(2)]
+
+    reference = te.Linear(*shape, bias=False, params_dtype=torch.bfloat16, device="cuda")
+    fused = te.Linear(*shape, bias=False, params_dtype=torch.bfloat16, device="cuda")
+    with torch.no_grad():
+        fused.weight.copy_(reference.weight)
+
+    accumulated = torch.zeros(shape, device="cuda", dtype=torch.float32)
+    for x in inputs:
+        reference(x).sum().backward()
+        accumulated.add_(reference.weight.grad.data)
+        reference.weight.grad = None
+
+    fused.weight.main_grad = torch.zeros(shape, device="cuda", dtype=torch.float32)
+    fused.weight.grad_added_to_main_grad = False
+    chunk = torch.nn.Module()
+    chunk.fc = fused
+    assert _enable_wgrad_accumulation_fusion([chunk]) == 1
+    for x in inputs:
+        fused(x).sum().backward()
+
+    assert fused.weight.main_grad.abs().max() > 0, "fusion produced no gradient at all"
+    torch.testing.assert_close(fused.weight.main_grad, accumulated, rtol=2e-2, atol=2e-2)
